Account Creation page cleaned up and added validations

+++ trunk/src/main/java/com/openbank/onlinebanking/controller/staff/StaffController.java
@@ -1,10 +1,16 @@
 package com.openbank.onlinebanking.controller.staff;
 
 import java.util.Date;
+import java.util.LinkedHashMap;
+import java.util.Map;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Controller;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.Errors;
+import org.springframework.validation.ObjectError;
+import org.springframework.validation.ValidationUtils;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -19,7 +25,10 @@ import com.openbank.onlinebanking.dto.Profile;
 import com.openbank.onlinebanking.dto.Role;
 import com.openbank.onlinebanking.dto.User;
 import com.openbank.onlinebanking.form.staff.CreateAccountForm;
+import com.openbank.onlinebanking.util.AppUtil;
+
 import static com.openbank.onlinebanking.util.ApplicationConstants.CUSTOMER_ROLE;
+import static com.openbank.onlinebanking.util.ApplicationConstants.DATE_FORMAT;
 
 @Controller
 public class StaffController {
@@ -38,50 +47,72 @@ public class StaffController {
 		createAccountForm.setTenantId(tenantId);
 		createAccountForm.setStaffProfileId(profileId);
 		ModelAndView modelAndView = new ModelAndView("staffcreateaccount");
+		modelAndView.addObject("accountTypeList", loadAccountTypeMap());
+		
 		modelAndView.addObject("form", createAccountForm);
 		log.debug("Existing..........");
 		return modelAndView;
 	}
 
 	@RequestMapping(value="/staffcreateaccount",  method=RequestMethod.POST)
-	public ModelAndView createAccountFormSubmit(@ModelAttribute CreateAccountForm createAccountForm ) {
+	public ModelAndView createAccountFormSubmit(@ModelAttribute CreateAccountForm createAccountForm, BindingResult result) {
 
 		log.debug("Entering - CreateAccountForm : {}", createAccountForm.toString());
 		ModelAndView modelAndView = new ModelAndView("staffcreateaccount");
-
-		String tenantId = createAccountForm.getTenantId();
-		Profile staffProfile = profileService.getProfileById(createAccountForm.getStaffProfileId(), tenantId);
-		if(staffProfile != null) {
-			createAccountForm.setStaffFirstName(staffProfile.getFirstName());
-			createAccountForm.setStaffLastName(staffProfile.getLastName());
+		boolean isSuccess = true;
+		validate(createAccountForm, result);
+		if(createAccountForm.getPhoneNo()!= null) {
+			if(!AppUtil.isAValidDDMMYYYYDate(createAccountForm.getDateOfBirth())) {
+				log.error("Invalid date format " + createAccountForm.getDateOfBirth()); 
+				isSuccess = false;
+				result.addError(new ObjectError("dateOfBirth", "Invalid dateOfBirth. Please use the format " + DATE_FORMAT ));
+			}
 		}
-		
-		
-		//Create Profile
-		String profileId = profileService.saveProfile(createProfile(createAccountForm));
-		String accountNo = null;
-		try {
-			//Create Use Account
-			loginService.CreateUser(createUser(createAccountForm, profileId));
-
-		} catch (Exception e) {
-			log.error("Error Creating User {} ", e);
-			profileService.deleteProfile(profileId, tenantId);
-		}
-		
-		try {
-			//Create Account
-			accountNo = accountService.saveAccount(createAccount(createAccountForm, profileId));
-			log.debug("Account No. " + accountNo);
-		} catch (Exception e) {
-			log.error("Error Creating Account {} ", e);
-			loginService.deleteUser(createAccountForm.getUserName(), tenantId);
-			profileService.deleteProfile(profileId, tenantId);
+		if(isSuccess && !result.hasErrors()) {
 			
+			String tenantId = createAccountForm.getTenantId();
+			Profile staffProfile = profileService.getProfileById(createAccountForm.getStaffProfileId(), tenantId);
+			if(staffProfile != null) {
+				createAccountForm.setStaffFirstName(staffProfile.getFirstName());
+				createAccountForm.setStaffLastName(staffProfile.getLastName());
+			}
+			
+			
+			//Create Profile
+			String profileId = profileService.saveProfile(createProfile(createAccountForm));
+			String accountNo = null;
+			try {
+				//Create Use Account
+				loginService.CreateUser(createUser(createAccountForm, profileId));
+
+			} catch (Exception e) {
+				log.error("Error Creating User {} ", e);
+				isSuccess = false;
+				profileService.deleteProfile(profileId, tenantId);
+				
+			}
+			
+			try {
+				//Create Account
+				accountNo = accountService.saveAccount(createAccount(createAccountForm, profileId));
+				log.debug("Account No. " + accountNo);
+			} catch (Exception e) {
+				log.error("Error Creating Account {} ", e);
+				isSuccess = false;
+				loginService.deleteUser(createAccountForm.getUserName(), tenantId);
+				profileService.deleteProfile(profileId, tenantId);
+				
+			}
+			if(isSuccess) {
+				modelAndView.addObject("successMessage", "Account '"+accountNo+"' successfully created !!!");
+				resetForm(createAccountForm);
+			} else {
+				result.addError(new ObjectError("account", "Error Opening new account"));
+			}
+
 		}
-		modelAndView.addObject("successMessage", "Account '"+accountNo+"' successfully created !!!");
-		resetForm(createAccountForm);
 		modelAndView.addObject("form", createAccountForm);
+		modelAndView.addObject("accountTypeList", loadAccountTypeMap());
 		log.debug("Existing..........");
 		return modelAndView;
 	}
@@ -94,9 +125,31 @@ public class StaffController {
 		createAccountForm.setBranchCode(null);
 		createAccountForm.setSex(null);
 		createAccountForm.setUserName(null);
-		
-		
+		createAccountForm.setAddress(null);
+		createAccountForm.setPhoneNo(null);
+		createAccountForm.setEmailAddress(null);
 	}
+	
+	public void validate(Object target, Errors errors) {
+		ValidationUtils.rejectIfEmptyOrWhitespace(errors, "accountType", "","Account Type cannot be blank");
+	    ValidationUtils.rejectIfEmptyOrWhitespace(errors, "firstName", "", "First Name cannot be blank");
+	    ValidationUtils.rejectIfEmptyOrWhitespace(errors, "lastName", "", "Last Name cannot be blank");
+	    ValidationUtils.rejectIfEmptyOrWhitespace(errors, "sex", "", "Sex cannot be blank");
+	    ValidationUtils.rejectIfEmptyOrWhitespace(errors, "dateOfBirth", "", "Date of Birth cannot be blank");
+	    ValidationUtils.rejectIfEmptyOrWhitespace(errors, "userName", "", "User Name cannot be blank");
+	    ValidationUtils.rejectIfEmptyOrWhitespace(errors, "address", "", "Address cannot be blank");
+	    ValidationUtils.rejectIfEmptyOrWhitespace(errors, "emailAddress", "", "Email Address cannot be blank");
+	    ValidationUtils.rejectIfEmptyOrWhitespace(errors, "phoneNo", "", "Phone cannot be blank");
+	    
+	}
+	
+	
+	private Map<String,String> loadAccountTypeMap() {
+		Map<String,String> modeMap = new LinkedHashMap<String,String>();
+			modeMap.put("Savings", "Savings");
+			modeMap.put("Current", "Current");
+		return modeMap;
+	}	
 
 	private User createUser(CreateAccountForm createAccountForm, String profileId) {
 		User user = new User();
@@ -115,7 +168,10 @@ public class StaffController {
 		Profile profile = new Profile();
 		profile.setFirstName(createAccountForm.getFirstName());
 		profile.setLastName(createAccountForm.getLastName());
-		//profile.setDateOfBirth(createAccountForm.getDateOfBirth());  //FIX ME
+		profile.setDateOfBirth(AppUtil.getDateFromStr(createAccountForm.getDateOfBirth()));
+		profile.setAddress(createAccountForm.getAddress());
+		profile.setEmailAddress(createAccountForm.getEmailAddress());
+		profile.setPhone(createAccountForm.getPhoneNo());
 		profile.setTenantId(createAccountForm.getTenantId());
 		profile.setSex(createAccountForm.getSex());
 		profile.setCreatedBy(createAccountForm.getStaffFirstName() +" "+ createAccountForm.getStaffLastName());
